webapp/parkinsons-voice-mvp/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
numpy==2.1.3
scipy==1.14.1
librosa==0.10.2.post1
soundfile==0.12.1
praat-parselmouth==0.4.5
pydantic==2.11.7
